requirements.txt
langchain
openai
playwright
beautifulsoup4
streamlit
python-dotenv